requirements.txt
# Core libraries for training and inference
torch
transformers
datasets
peft
sentence-transformers

# Utility libraries
numpy
scikit-learn
matplotlib
psutil
tqdm
pyyaml
ftfy
librosa 
scipy
soundfile
kokoro
nltk
keybert

# PDF processing
pdfminer.six
PyMuPDF

# Memory and embeddings
faiss-cpu # or faiss-gpu

# Optional interactive/monitoring tools
wandb
spacy
gradio
requests
